mcp-s3 v1.4.0 and mcp-trino v1.4.0 carry the upstream fixes, so the S3 schema override goes
mcp-s3 v1.4.0 admits null on every slice-backed output property (txn2/mcp-s3#141), so pkg/toolkits/s3/output_schemas.go and its WithOutputSchemas call are removed; the failed-listing proof stays as a pin on the upstream fix. mcp-trino v1.4.0 declares open output schemas explicitly (txn2/mcp-trino#85); MCPOutputSchemaMiddleware stays, since the keys it admits are the platform's own and apply to every advertised schema.
