+++ docs/llms-full.txt
@@ -1190,7 +1190,7 @@ ResultType → Icons → DescOverrides → Visibility → OutputSchema → Apps 
 
 For `tools/call` requests: Auth creates the `PlatformContext` (user identity, persona, toolkit metadata) and records tool calls for workflow tracking. Audit reads it to build the audit event. Rules checks session workflow state and prepends warnings when needed. Enrichment adds cross-service context and optional discovery notes. Unauthorized requests are rejected before reaching audit, so only authenticated tool calls appear in audit logs. ResultType, the outermost layer, types every `tools/call`, `prompts/get` and `resources/read` result with the `resultType` the negotiated protocol revision (`2026-07-28` and later) requires: the SDK stamps only the result its own handler returns, so a gate's refusal, the error contract's rebuild of a bare error, and a managed resource read, all built by the layers below, are typed here; an older client receives the field unset, as the SDK leaves it.
 
-For `tools/list` responses: Icons, DescOverrides, and Visibility modify tool metadata and filter the tool list, and OutputSchema opens the top level of every advertised output schema (`additionalProperties` allowed, nothing required, the error envelope documented under `error`) so a client validating a result against the schema it was advertised accepts the keys the chain adds to `structuredContent` — the error envelope, `call_reference`, enrichment. A toolkit registering typed handlers without an explicit schema (mcp-trino) otherwise advertises the closed schema the SDK infers from its struct; the server still validates the handler's own output against that inferred schema before any middleware runs.
+For `tools/list` responses: Icons, DescOverrides, and Visibility modify tool metadata and filter the tool list, and OutputSchema opens the top level of every advertised output schema (`additionalProperties` allowed, nothing required, the error envelope documented under `error`) so a client validating a result against the schema it was advertised accepts the keys the chain adds to `structuredContent` — the error envelope, `call_reference`, enrichment. A toolkit registering typed handlers without an explicit schema (mcp-trino before v1.4.0) otherwise advertises the closed schema the SDK infers from its struct; the server still validates the handler's own output against that inferred schema before any middleware runs.
 
 Audit delivery is configurable (`audit.delivery`). In the default **async** mode, events are enqueued on a bounded writer drained by a single background goroutine, so store latency never blocks a tool response, and a graceful shutdown drains that queue (bounded by a 10s deadline) before the store and database are closed; under sustained backpressure or a crash, queued events are dropped and counted by the `audit_events_dropped_total` metric. In **sync** mode, the event is written on the request goroutine with a per-write timeout, trading latency for backpressure and zero queue-overflow drops. In either mode, a write that fails is logged via `slog.Error` but the tool call still succeeds.
