+++ docs/reference/middleware.md
@@ -193,7 +193,7 @@ Opens the top level of every output schema a `tools/list` response advertises (i
 func MCPOutputSchemaMiddleware() mcp.Middleware
 ```
 
-A toolkit that registers a typed handler with no explicit output schema (mcp-trino does) gets one inferred from its Go struct, and jsonschema-go closes every struct-derived object with `additionalProperties: false` and a `required` list; a client validating against that schema discarded every Trino result. The decorator gives each advertised object schema the same contract `middleware.OpenToolOutputSchema` gives the platform-owned tools: `additionalProperties` allowed, nothing required, and the error envelope documented under `error`. Nested schemas are untouched, the server's registry is never mutated (the listed `Tool` is replaced with a copy), and the SDK still validates a handler's own structured output against the schema it inferred, inside the handler wrapper and before any middleware runs. A tool that declares a non-object schema, or none, is listed as it is.
+A toolkit that registers a typed handler with no explicit output schema (mcp-trino did before v1.4.0) gets one inferred from its Go struct, and jsonschema-go closes every struct-derived object with `additionalProperties: false` and a `required` list; a client validating against that schema discarded every Trino result. The decorator gives each advertised object schema the same contract `middleware.OpenToolOutputSchema` gives the platform-owned tools: `additionalProperties` allowed, nothing required, and the error envelope documented under `error`. Nested schemas are untouched, the server's registry is never mutated (the listed `Tool` is replaced with a copy), and the SDK still validates a handler's own structured output against the schema it inferred, inside the handler wrapper and before any middleware runs. A tool that declares a non-object schema, or none, is listed as it is.
 
 ### MCPSessionHandleSchemaMiddleware
